dev-requirements-local.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=dev-requirements-local.txt dev-requirements.in
#
alabaster==0.7.12         # via sphinx
argparse==1.4.0           # via -c requirements.txt, unittest2
attrs==19.3.0             # via jsonschema
aws-sam-translator==1.22.0  # via cfn-lint
aws-xray-sdk==2.5.0       # via moto
babel==2.3.4              # via -c requirements.txt, sphinx
backports.ssl-match-hostname==3.7.0.1  # via docker
backports.tempfile==1.0   # via moto
backports.weakref==1.0.post1  # via backports.tempfile
beautifulsoup4==4.5.1     # via -r dev-requirements.in
blinker==1.4              # via flask-debugtoolbar
boto3==1.12.43            # via aws-sam-translator, moto
boto==2.49.0              # via -c requirements.txt, moto
botocore==1.15.43         # via aws-xray-sdk, boto3, moto, s3transfer
certifi==2020.4.5.1       # via -c requirements.txt, requests
cffi==1.14.0              # via cryptography
cfn-lint==0.29.6          # via moto
chardet==3.0.4            # via -c requirements.txt, requests
click==7.1.1              # via -c requirements.txt, flask
configparser==4.0.2       # via importlib-metadata
contextlib2==0.6.0.post1  # via importlib-metadata, importlib-resources, zipp
cookies==2.2.1            # via responses
cryptography==2.9.1       # via moto, sshpubkeys
decorator==4.4.2          # via -c requirements.txt, networkx
docker==4.2.0             # via moto
docutils==0.12            # via -r dev-requirements.in, botocore, sphinx
ecdsa==0.15               # via python-jose, sshpubkeys
enum34==1.1.10            # via aws-sam-translator, aws-xray-sdk, cryptography
extras==1.0.0             # via testtools
factory-boy==2.1.1        # via -r dev-requirements.in
fixtures==3.0.0           # via mox3, testtools
flask-debugtoolbar==0.10.1  # via -r dev-requirements.in
flask==0.12.4             # via -c requirements.txt, flask-debugtoolbar
funcsigs==1.0.2           # via -c requirements.txt, mock
functools32==3.2.3.post2  # via jsonschema
future==0.18.2            # via aws-xray-sdk
futures==3.3.0            # via s3transfer
idna==2.8                 # via -c requirements.txt, moto, requests
imagesize==1.2.0          # via sphinx
importlib-metadata==1.6.0  # via importlib-resources, jsonpickle, jsonschema
importlib-resources==1.4.0  # via cfn-lint
ipaddress==1.0.23         # via -c requirements.txt, cryptography, docker
itsdangerous==1.1.0       # via -c requirements.txt, flask, flask-debugtoolbar
jinja2==2.11.2            # via -c requirements.txt, flask, moto, sphinx
jmespath==0.9.5           # via boto3, botocore
jsondiff==1.1.2           # via moto
jsonpatch==1.25           # via cfn-lint
jsonpickle==1.4.1         # via aws-xray-sdk
jsonpointer==2.0          # via jsonpatch
jsonschema==3.2.0         # via aws-sam-translator, cfn-lint
linecache2==1.0.0         # via traceback2
markupsafe==1.1.1         # via -c requirements.txt, jinja2
mock==2.0.0               # via -r dev-requirements.in, moto, responses
moto==1.3.14              # via -r dev-requirements.in
mox3==0.28.0              # via pyfakefs
networkx==2.2             # via cfn-lint
packaging==20.3           # via sphinx
pathlib2==2.3.5           # via cfn-lint, importlib-metadata, importlib-resources
pbr==5.4.5                # via -c requirements.txt, fixtures, mock, mox3, testtools
pyasn1==0.4.8             # via python-jose, rsa
pycodestyle==2.2.0        # via -r dev-requirements.in
pycparser==2.20           # via cffi
pyfakefs==2.9             # via -r dev-requirements.in
pygments==2.5.2           # via -c requirements.txt, sphinx
pyparsing==2.4.7          # via -c requirements.txt, packaging
pyrsistent==0.16.0        # via jsonschema
python-dateutil==2.8.1    # via -c requirements.txt, botocore, moto
python-jose==3.1.0        # via moto
python-mimeparse==1.6.0   # via testtools
pytz==2016.7              # via -c requirements.txt, babel, moto
pyyaml==5.2               # via -c requirements.txt, cfn-lint, moto
requests==2.23.0          # via -c requirements.txt, docker, moto, responses, sphinx
responses==0.10.14        # via moto
rsa==4.0                  # via python-jose
s3transfer==0.3.3         # via boto3
scandir==1.10.0           # via pathlib2
singledispatch==3.4.0.3   # via importlib-resources
six==1.14.0               # via -c requirements.txt, aws-sam-translator, cfn-lint, cryptography, docker, ecdsa, fixtures, jsonschema, mock, moto, packaging, pathlib2, pyrsistent, python-dateutil, python-jose, responses, singledispatch, sphinx, testtools, unittest2, websocket-client
snowballstemmer==2.0.0    # via sphinx
sphinx-rtd-theme==0.3.1   # via -r dev-requirements.in
sphinx==1.7.1             # via -r dev-requirements.in
sphinxcontrib-websupport==1.1.2  # via sphinx
sshpubkeys==3.1.0         # via moto
testtools==2.4.0          # via fixtures
traceback2==1.4.0         # via testtools, unittest2
typing==3.7.4.1           # via importlib-resources, sphinx
unittest2==1.1.0          # via testtools
urllib3==1.25.9           # via -c requirements.txt, botocore, requests
websocket-client==0.57.0  # via docker
werkzeug==0.16.1          # via -c requirements.txt, flask, flask-debugtoolbar, moto
wrapt==1.12.1             # via aws-xray-sdk
xmltodict==0.12.0         # via moto
zipp==1.2.0               # via importlib-metadata, importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
